requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
ciso8601==2.1.3
Django==3.0.3
django-crispy-forms==1.9.0
django-recaptcha==2.0.6
idna==2.9
json2parquet==0.0.28
numpy==1.18.1
pandas==0.24.2
pyarrow==0.16.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
